UDP packet data pointed at wrong location
I have been working on implementing a port to the atmel
ATMega128 and I have noticed that sending UDP packets does not
work (the transferred data is always 0).

A look into the code reveals that in uip.c, the address of
uip_appdata gets set incorrectly.

Somewhere around line 1159 the code needs to be changed:

--- uip_appdata = &uip_buf[UIP_LLH_LEN + UIP_IPTCPH_LEN];
+++ uip_appdata = &uip_buf[UIP_LLH_LEN + UIP_IPUDPH_LEN];

http://article.gmane.org/gmane.network.uip.user/68